backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
requests
beautifulsoup4
